// src/app/core/models/catalogs/cuenta.model.ts
export interface ICuenta {
  accountNumber?: number;
  accountNumberTransfer: string;
  registerNumber?: string;
  delegationNumber: number;
  cveAccount: string;
  accountType: string;
  cveCurrency: string;
  square: null;
  branch: number;
  cveInterestCalcRate: string;
  cveBank: string;
  isReference?: boolean;
}